Version 2
The alignment for the candle flames being off always bugged me when I remembered this project so even though it's like 2 years later and nobody is ever gonna use this again I decided to fix it. Everything's properly aligned now with flexbox instead of the random VW and VH that I was guessing at before. Plus I rewrote the JS logic so it's not as needlessly repetitive and long